requirements.txt
# Core dependencies
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0

# Model architectures
efficientnet-pytorch

# Image processing
Pillow>=10.0.0
opencv-python

# Audio processing
librosa>=0.10.0
soundfile

# Utilities
numpy>=1.24.0
scipy>=1.11.0